+++ Interface/AddOns/NeatPlatesHub/Defaults.lua
@@ -294,6 +294,15 @@ NeatPlatesHubDefaults = {
 	WidgetThreatPercentage = false,
 	WidgetArenaIcon = false,
 
+	-- Resource Widget
+	---------------------------------------
+	WidgetResourceMode = 1,
+	WidgetResourceSpacing = 0,
+	WidgetResourceStyle = 'Neat',
+	WidgetResourceTimerFontSize = 8,
+	WidgetResourceDisplayTimer = true,
+	WidgetResourceHideEmpty = false,
+
 	-- Aura Widget
 	---------------------------------------
 	WidgetDebuff = true,