+++ Interface/AddOns/NeatPlates/CHANGELOG.md
@@ -1,6 +1,8 @@
 # NeatPlates
 
-## [v414](https://github.com/Luxocracy/NeatPlates/tree/v414) (2022-09-02)
-[Full Changelog](https://github.com/Luxocracy/NeatPlates/compare/V413...v414) [Previous Releases](https://github.com/Luxocracy/NeatPlates/releases)
+## [v427](https://github.com/Luxocracy/NeatPlates/tree/v427) (2022-11-16)
+[Full Changelog](https://github.com/Luxocracy/NeatPlates/compare/v426...v427) [Previous Releases](https://github.com/Luxocracy/NeatPlates/releases)
 
-- Display combo points for rogues even if they have no points, as long as they have a charged combo point  
+- Locale update  
+- Fix Evoker blizzlike art  
+- Add Evoker to resource widget